pyproject.toml
[tool.poetry]
name = "parallel-trading"
version = "0.1.0"
description = ""
authors = ["diamondrigido <dimasik1998.98@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
SQLAlchemy = "^1.4.27"
pandas = "^1.3.4"
python-binance = "^1.0.15"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
